requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-telegram-bot==20.7
google-generativeai>=0.8.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
aiohttp==3.9.1
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0
pdf2image==1.16.3
pytesseract==0.3.10


